Move on refactoring
