+++ argopy/stores/index/implementations/index_pyarrow.py
@@ -15,17 +15,17 @@ try:
 except ModuleNotFoundError:
     pass
 
-from ...errors import DataNotFound, InvalidDatasetStructure
-from ...utils.checkers import check_index_cols, is_indexbox, check_wmo, check_cyc
-from ...utils.casting import to_list
-from .argo_index_proto import ArgoIndexStoreProto
-from .argo_index_proto_s3 import search_s3
+from ....errors import DataNotFound, InvalidDatasetStructure
+from ....utils.checkers import check_index_cols, is_indexbox, check_wmo, check_cyc
+from ....utils.casting import to_list
+from ..spec import ArgoIndexStoreProto
+from .index_s3 import search_s3
 
 
 log = logging.getLogger("argopy.stores.index.pa")
 
 
-class indexstore_pyarrow(ArgoIndexStoreProto):
+class indexstore(ArgoIndexStoreProto):
     """Argo GDAC index store using :class:`pyarrow.Table` as internal storage format.
 
     With this store, index and search results are saved as pyarrow/parquet files in cache
