+++ argopy/stores/index/spec.py
@@ -21,7 +21,7 @@ from ...errors import GdacPathError, S3PathError, InvalidDataset, OptionValueErr
 from ...utils.checkers import isconnected, has_aws_credentials
 from ...utils.accessories import Registry
 from .. import httpstore, memorystore, filestore, ftpstore, s3store
-from .argo_index_proto_s3 import get_a_s3index
+from .implementations.index_s3 import get_a_s3index
 
 try:
     import pyarrow.csv as csv  # noqa: F401
@@ -296,11 +296,12 @@ class ArgoIndexStoreProto(ABC):
 
     @property
     def index_path(self):
+        """Absolute path to the index file"""
         return self.fs["src"].fs.sep.join([self.host, self.index_file])
 
     @property
     def cname(self) -> str:
-        """Return the search constraint(s) as a pretty formatted string
+        """Search constraint(s) as a pretty formatted string
 
         Return 'full' if a search was not yet performed on the indexstore instance
 
@@ -624,7 +625,7 @@ class ArgoIndexStoreProto(ABC):
         else:
             log.debug("Converting [%s] to dataframe from scratch ..." % src)
             # Post-processing for user:
-            from ..related import load_dict, mapp_dict
+            from ...related import load_dict, mapp_dict
 
             if nrows is not None:
                 df = df.loc[0 : nrows - 1].copy()
