+++ argopy/tests/test_stores_index.py
@@ -17,7 +17,7 @@ from argopy.errors import (
     InvalidDatasetStructure,
 )
 from argopy.utils.checkers import is_list_of_strings
-from argopy.stores.argo_index_pd import indexstore_pandas
+from argopy.stores import indexstore_pd
 from mocked_http import mocked_httpserver, mocked_server_address
 
 log = logging.getLogger("argopy.tests.indexstores")
@@ -471,7 +471,7 @@ class IndexStore_test_proto:
 @skip_CORE
 class Test_IndexStore_pandas_CORE(IndexStore_test_proto):
     network = "core"
-    indexstore = indexstore_pandas
+    indexstore = indexstore_pd
     index_file = "ar_index_global_prof.txt"
 
 
@@ -479,7 +479,7 @@ class Test_IndexStore_pandas_CORE(IndexStore_test_proto):
 @skip_BGCs
 class Test_IndexStore_pandas_BGC_synthetic(IndexStore_test_proto):
     network = "bgc"
-    indexstore = indexstore_pandas
+    indexstore = indexstore_pd
     index_file = "argo_synthetic-profile_index.txt"
 
 
@@ -487,7 +487,7 @@ class Test_IndexStore_pandas_BGC_synthetic(IndexStore_test_proto):
 @skip_BGCb
 class Test_IndexStore_pandas_BGC_bio(IndexStore_test_proto):
     network = "bgc"
-    indexstore = indexstore_pandas
+    indexstore = indexstore_pd
     index_file = "argo_bio-profile_index.txt"
 
 
@@ -500,9 +500,9 @@ class Test_IndexStore_pandas_BGC_bio(IndexStore_test_proto):
 @skip_CORE
 class Test_IndexStore_pyarrow_CORE(IndexStore_test_proto):
     network = "core"
-    from argopy.stores.argo_index_pa import indexstore_pyarrow
+    from argopy.stores import indexstore_pa
 
-    indexstore = indexstore_pyarrow
+    indexstore = indexstore_pa
     index_file = "ar_index_global_prof.txt"
 
 
@@ -511,9 +511,9 @@ class Test_IndexStore_pyarrow_CORE(IndexStore_test_proto):
 @skip_BGCs
 class Test_IndexStore_pyarrow_BGC_bio(IndexStore_test_proto):
     network = "bgc"
-    from argopy.stores.argo_index_pa import indexstore_pyarrow
+    from argopy.stores import indexstore_pa
 
-    indexstore = indexstore_pyarrow
+    indexstore = indexstore_pa
     index_file = "argo_bio-profile_index.txt"
 
 
@@ -522,7 +522,7 @@ class Test_IndexStore_pyarrow_BGC_bio(IndexStore_test_proto):
 @skip_BGCb
 class Test_IndexStore_pyarrow_BGC_synthetic(IndexStore_test_proto):
     network = "bgc"
-    from argopy.stores.argo_index_pa import indexstore_pyarrow
+    from argopy.stores import indexstore_pa
 
-    indexstore = indexstore_pyarrow
+    indexstore = indexstore_pa
     index_file = "argo_synthetic-profile_index.txt"
