+++ docs/redirects.txt
@@ -13,3 +13,5 @@ generated/argopy.utilities.ArgoNVSReferenceTables.rst generated/argopy.related.A
 # generated/argopy.utilities.load_dict generated/argopy.related.load_dict.rst
 generated/argopy.xarray.ArgoAccessor.transform_data_mode.rst generated/argopy.Dataset.argo.datamode.merge.rst
 generated/argopy.xarray.ArgoAccessor.filter_data_mode.rst generated/argopy.Dataset.argo.datamode.filter.rst
+#
+
